* [jsfm] add debug utils
